ReflectionUtil.getDomainClasses() now returns a Set<Class>
ReflectionUtilTest.testGetAllClasses() was failing because, on my setup, classLoader.getResources(""); returned the
directory were my classes were held twice.  This meant that getDomainClasses() returned 92 classes, because the
expected 46 were all doubled up.  A Set seems like a more appropriate data structure in this case, anyway, and
was a useful way to remove duplicates.